+++ src/main/java/domain/LadderGame/LadderGame.java
@@ -5,6 +5,7 @@ import domain.Collection.Participants;
 import domain.Collection.Result;
 import domain.Collection.Results;
 import domain.Ladder.Ladder;
+import domain.util.SequenceSwapper;
 
 import java.util.HashMap;
 import java.util.List;
